improved analysis module structure
