+++ analysis/build.gradle
@@ -19,6 +19,7 @@ android {
 
     aaptOptions {
         noCompress "tflite", "csv"
+
     }
 
     buildTypes {
