+++ analysis/src/main/AndroidManifest.xml
