+++ app/build.gradle
@@ -22,6 +22,16 @@ android {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+
+            ndk {
+                abiFilters 'armeabi-v7a', 'arm64-v8a'
+            }
+        }
+
+        debug {
+            ndk {
+                abiFilters 'armeabi-v7a', "x86_64"
+            }
         }
     }
 
@@ -44,7 +54,6 @@ dependencies {
     implementation 'com.google.android.material:material:1.3.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     implementation 'androidx.appcompat:appcompat:1.3.0'
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
@@ -53,8 +62,6 @@ dependencies {
     implementation 'androidx.navigation:navigation-fragment-ktx:2.3.5'
     implementation 'androidx.navigation:navigation-ui-ktx:2.3.5'
 
-    //api
-    implementation project(":api")
 
     // CameraX core library using camera2 implementation
     implementation "androidx.camera:camera-camera2:1.0.0"
@@ -64,5 +71,7 @@ dependencies {
     implementation "androidx.camera:camera-view:1.0.0-alpha24"
 
 
-    implementation project(':classifier')
+    //api
+    implementation project(":api")
+    implementation project(':analysis')
 }
