+++ build.gradle
@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
     ext{
-        kotlin_version = "1.4.32"
+        kotlin_version = "1.5.0"
         retrofit_version = "2.9.0"
         moshi_version = "1.12.0"
     }
